fix(animation): 修复退场动画不渲染的问题 + 移除 AnimationDriver 死代码
核心问题：Modal/Drawer 的 close() 立即设 visible=false，
导致渲染管线/LayerTree/动画更新路径全部跳过该 widget，
退场动画永远不会被推进和绘制。

解决方案（无需修改 WidgetTree/LayerTree）：
- Modal/Drawer 新增  字段
- close() 不再设 visible=false，而是启动退场 TransitionPlayer
  并保持 visible=true，等动画结束后自动设 visible=false
- render()： 时才跳过（完全隐藏）
- preferred_size()：closing 期间保持布局空间
- on_event()：closing 期间不处理事件
- open()：若在退场中则取消退场直接显示

移除：
- AnimationDriver / AnimValue / AnimationCallback（零引用死代码）
- driver.rs 文件
- api/types.rs 和 animation/mod.rs 中相关导出/文档

+++ demo/src/demos/dashboard/widgets.rs
@@ -37,8 +37,8 @@ define_widget! {
         Size::new(48.0, 48.0)
     }
 
-    on_update => (&mut self, dt: f32) {
-        self.time += dt;
+    on_update => (&mut self, dt: f64) {
+        self.time += dt as f32;
         if self.time > std::f32::consts::TAU { self.time -= std::f32::consts::TAU; }
     }
 
@@ -80,8 +80,8 @@ define_widget! {
         Size::new(200.0, 60.0)
     }
 
-    on_update => (&mut self, dt: f32) {
-        self.time += dt;
+    on_update => (&mut self, dt: f64) {
+        self.time += dt as f32;
         if self.time > 2.0 { self.time -= 2.0; }
     }
